@@ -1,41 +1,59 @@
-const TypeDoc = require('typedoc');
-const { Project } = require('ts-morph');
-const { createRequire } = require('module');
-const { resolve: resolvePath, relative: relativePath } = require('path');
-const { existsSync, readFileSync, readdirSync, mkdirSync, writeFileSync, rmSync } = require('fs');
-const { split, replacePieces } = require('./split.js');
-const { execSync } = require('child_process');
-const { loadHooks } = require('./hooks.js');
-const { basePath, originalPath, translatingPath, translatedPath, distPath } = require('./utils.js');
+import { execSync } from 'child_process';
+import { existsSync, mkdirSync, readFileSync, readdirSync, rmSync, writeFileSync } from 'fs';
+import { createRequire } from 'module';
+import { relative as relativePath, resolve as resolvePath } from 'path';
+import { Project, SourceFile } from 'ts-morph';
+import type { PackageJson } from 'type-fest';
+import * as TypeDoc from 'typedoc';
+import runHooks from './hooks.js';
+import { replacePieces, split } from './split.js';
+import { basePath, distPath, originalPath, translatedPath, translatingPath } from './utils.js';
+
+declare module 'typedoc' {
+    interface TranslatableStrings {
+        tag_rc: [];
+    }
+}
 
 const namespacePrefix = '@minecraft/';
 const botModules = ['@minecraft/vanilla-data'];
-const skipResolutionModules = [];
+const skipResolutionModules: string[] = [];
 
-function readPackageInfo(modulePath) {
+function readPackageInfo(modulePath: string) {
     const packageInfoPath = resolvePath(modulePath, 'package.json');
     if (existsSync(packageInfoPath)) {
         try {
-            return JSON.parse(readFileSync(packageInfoPath, 'utf-8'));
-        } catch (e) {
+            return JSON.parse(readFileSync(packageInfoPath, 'utf-8')) as PackageJson;
+        } catch {
             /* ignore */
         }
     }
 }
 
-function findModule(moduleName, root) {
-    const localRequire = createRequire(resolvePath(root, 'node_modules'));
-    const searchPaths = localRequire.resolve.paths(moduleName);
-    for (const searchPath of searchPaths) {
-        const modulePath = resolvePath(searchPath, moduleName);
-        const moduleDesc = readPackageInfo(modulePath);
-        if (moduleDesc && moduleDesc.name === moduleName) {
-            return modulePath;
-        }
+function readPackageInfoOrThrow(modulePath: string) {
+    const packageInfo = readPackageInfo(modulePath);
+    if (!packageInfo) {
+        throw new Error(`package.json not exist or cannot read: ${modulePath}`);
     }
+    return packageInfo;
 }
 
-function walkFiles(directory, walker) {
+function findModuleOrThrow(moduleName: string, root: string) {
+    const localRequire = createRequire(resolvePath(root, 'node_modules'));
+    const searchPaths = localRequire.resolve.paths(moduleName);
+    if (searchPaths) {
+        for (const searchPath of searchPaths) {
+            const modulePath = resolvePath(searchPath, moduleName);
+            const moduleDesc = readPackageInfo(modulePath);
+            if (moduleDesc && moduleDesc.name === moduleName) {
+                return modulePath;
+            }
+        }
+    }
+    throw new Error(`Cannot find module ${moduleName} in ${root}`);
+}
+
+function walkFiles(directory: string, walker: (directory: string, fileName: string | null, path: string) => void) {
     const files = readdirSync(directory, { withFileTypes: true });
     walker(directory, null, directory);
     files.forEach((file) => {
@@ -47,26 +65,30 @@ function walkFiles(directory, walker) {
     });
 }
 
-function getModuleSourceFiles(fromPath, moduleSpecifier) {
+function getModuleSourceFiles(fromPath: string, moduleSpecifier: string) {
     const project = new Project();
     const sourceFile = project.createSourceFile(resolvePath(fromPath, '__temp_module_resolution__.ts'));
     const rootDecl = sourceFile.addExportDeclaration({ moduleSpecifier });
-    const referencedFiles = [];
-    const walk = (/** @type {import('ts-morph').SourceFile | undefined} */ source) => {
+    const referencedFiles: string[] = [];
+    const walk = (source: SourceFile | undefined) => {
         if (!source) return;
         const path = source.getFilePath();
         if (referencedFiles.includes(path)) return;
         referencedFiles.push(path);
         const importDecl = source.getImportDeclarations();
         const exportDecl = source.getExportDeclarations();
-        importDecl.forEach((decl) => walk(decl.getModuleSpecifierSourceFile()));
-        exportDecl.forEach((decl) => walk(decl.getModuleSpecifierSourceFile()));
+        importDecl.forEach((decl) => {
+            walk(decl.getModuleSpecifierSourceFile());
+        });
+        exportDecl.forEach((decl) => {
+            walk(decl.getModuleSpecifierSourceFile());
+        });
     };
     walk(rootDecl.getModuleSpecifierSourceFile());
     return referencedFiles.map((e) => resolvePath(e));
 }
 
-function getCommonStringFromStart(a, b) {
+function getCommonStringFromStart(a: string, b: string) {
     let len = Math.min(a.length, b.length);
     while (len > 0) {
         if (a.slice(0, len) === b.slice(0, len)) {
@@ -77,9 +99,8 @@ function getCommonStringFromStart(a, b) {
     return '';
 }
 
-async function build(translated) {
-    const runHooks = loadHooks();
-    const hookEventContext = { basePath, originalPath, translatingPath, translatedPath, distPath };
+export async function build(translated?: boolean) {
+    const hookContext = { basePath, originalPath, translatingPath, translatedPath, distPath };
 
     // 尝试加载翻译文件对应版本的 package.json
     console.time('[restoreDependencies] Total');
@@ -103,25 +124,24 @@ async function build(translated) {
 
     // 从依赖中构建用于生成文档的项目
     console.time('[loadOriginal] Total');
-    await runHooks('beforeLoad', hookEventContext);
+    await runHooks('beforeLoad', hookContext);
     const tsConfigFilePath = resolvePath(translatedPath, 'tsconfig.json');
     const project = new Project({
         tsConfigFilePath,
         skipAddingFilesFromTsConfig: true
     });
-    /** @type {import("ts-morph").SourceFile[]} */
-    const sourceFiles = [];
-    const dependencies = readPackageInfo(originalPath).dependencies;
+    const sourceFiles: SourceFile[] = [];
+    const dependencies = readPackageInfo(originalPath)?.dependencies ?? {};
     Object.keys(dependencies).forEach((moduleName) => {
         if (moduleName.startsWith(namespacePrefix)) {
             const pureModuleName = moduleName.slice(namespacePrefix.length);
-            const modulePath = findModule(moduleName, originalPath);
-            const packageInfo = readPackageInfo(modulePath);
+            const modulePath = findModuleOrThrow(moduleName, originalPath);
+            const packageInfo = readPackageInfoOrThrow(modulePath);
             const version = packageInfo.version;
-            console.log(`Loading d.ts for ${moduleName}@${version}`);
-            let dtsFiles = [];
+            console.log(`Loading d.ts for ${moduleName}@${version ?? 'undefined'}`);
+            let dtsFiles: string[] = [];
             walkFiles(modulePath, (dir, file, path) => {
-                if (file && file.endsWith('.d.ts')) {
+                if (file?.endsWith('.d.ts')) {
                     const relPath = relativePath(modulePath, path);
                     if (!relPath.includes('node_modules')) {
                         dtsFiles.push(path);
@@ -143,7 +163,7 @@ async function build(translated) {
                 );
                 if (!botModules.includes(moduleName)) sourceFiles.push(sourceFile);
             } else {
-                const typeEntry = resolvePath(modulePath, packageInfo.types).replace(/\.d\.ts$/i, '');
+                const typeEntry = resolvePath(modulePath, packageInfo.types ?? 'index.d.ts').replace(/\.d\.ts$/i, '');
                 const commonParent = dtsFiles
                     .map((path) => resolvePath(path, '..'))
                     .reduce((common, parent) => getCommonStringFromStart(common, parent));
@@ -171,8 +191,8 @@ async function build(translated) {
             dependencies[moduleName] = version;
         }
     });
-    Object.assign(hookEventContext, { basePath, project, sourceFiles, dependencies });
-    await runHooks('afterLoad', hookEventContext);
+    const translateHookContext = { ...hookContext, basePath, project, sourceFiles, dependencies };
+    await runHooks('afterLoad', translateHookContext);
     console.timeEnd('[loadOriginal] Total');
 
     if (translated) {
@@ -182,7 +202,7 @@ async function build(translated) {
             const pieces = split(sourceFile);
             replacePieces(sourceFile, pieces);
         });
-        await runHooks('afterTranslate', hookEventContext);
+        await runHooks('afterTranslate', translateHookContext);
         console.timeEnd('[translate] Total');
     }
 
@@ -206,35 +226,20 @@ async function build(translated) {
         true
     );
     rmSync(distPath, { recursive: true, force: true });
-    Object.assign(hookEventContext, { tsdocApplication });
-    await runHooks('beforeConvert', hookEventContext);
+    const beforeConvertContext = { ...translateHookContext, tsdocApplication };
+    await runHooks('beforeConvert', beforeConvertContext);
     const tsdocProject = await tsdocApplication.convert();
     console.timeEnd('[analyze] Total');
     if (tsdocProject) {
         console.time('[emit] Total');
-        Object.assign(hookEventContext, { tsdocProject });
-        await runHooks('afterConvert', hookEventContext);
+        const afterConvertContext = { ...beforeConvertContext, tsdocProject };
+        await runHooks('afterConvert', afterConvertContext);
         await tsdocApplication.generateDocs(tsdocProject, distPath);
         await tsdocApplication.generateJson(tsdocProject, resolvePath(distPath, 'index.json'));
-        await runHooks('afterEmit', hookEventContext);
+        await runHooks('afterEmit', afterConvertContext);
         console.timeEnd('[emit] Total');
+        return afterConvertContext;
     } else {
         throw new Error('Convert failed');
     }
-    return hookEventContext;
-}
-
-module.exports = {
-    build
-};
-
-if (require.main === module) {
-    build(true)
-        .then(() => {
-            process.exit(0);
-        })
-        .catch((err) => {
-            console.error(err);
-            process.exit(1);
-        });
 }
