@@ -1,8 +1,16 @@
-const { createHash } = require('crypto');
-const { mkdirSync, writeFileSync, existsSync, readFileSync } = require('fs');
-const { resolve: resolvePath } = require('path');
-const { SyntaxKind } = require('ts-morph');
-const { DocumentReflection, JSX } = require('typedoc');
+import { createHash } from 'crypto';
+import { existsSync, mkdirSync, readFileSync, writeFileSync } from 'fs';
+import { resolve as resolvePath } from 'path';
+import { Symbol, SyntaxKind, ts } from 'ts-morph';
+import {
+    DefaultTheme,
+    DocumentReflection,
+    JSX,
+    Reflection,
+    ReflectionSymbolId,
+    type CommentDisplayPart
+} from 'typedoc';
+import type { Hook } from './hook.js';
 
 const ExampleNameOverwrite = [
     {
@@ -19,45 +27,53 @@ const ExampleNameOverwrite = [
     }
 ];
 
-function hashTextShort(str) {
+function hashTextShort(str: string) {
     return createHash('sha256').update(str).digest('hex').slice(0, 8);
 }
 
-/**
- * @param {JSX.Children} jsx
- * @param {(element: JSX.Children, traversal: () => void) => void} f
- */
-function traversalJSX(jsx, f) {
+type TraversableJSXChildren = Exclude<JSX.Children, JSX.Children[] | null | undefined>;
+
+function traverseJSX(jsx: JSX.Children, f: (element: TraversableJSXChildren, traverseInto: () => void) => void) {
     if (Array.isArray(jsx)) {
         for (const child of jsx) {
-            traversalJSX(child, f);
+            traverseJSX(child, f);
         }
-    } else if (jsx !== null || jsx !== undefined) {
+    } else if (jsx !== null && jsx !== undefined) {
         f(jsx, () => {
             if (typeof jsx === 'object') {
                 for (const child of jsx.children) {
-                    traversalJSX(child, f);
+                    traverseJSX(child, f);
                 }
             }
         });
     }
 }
 
-function findJSXElement(jsx, predicate) {
-    const elements = [];
-    traversalJSX(jsx, (el, traversal) => {
+function findJSXElement<E extends TraversableJSXChildren>(
+    jsx: JSX.Children,
+    predicate: (element: TraversableJSXChildren) => element is E
+) {
+    const elements: E[] = [];
+    traverseJSX(jsx, (el, traverseInto) => {
         if (predicate(el)) {
             elements.push(el);
         }
-        traversal();
+        traverseInto();
     });
     return elements;
 }
 
-const examples = {};
+const examples: Record<
+    string,
+    {
+        content: string;
+        hash: string;
+        fileName: string;
+        sources: { source: string; fileName: string; path: string; symbol: Symbol }[];
+    }[]
+> = {};
 
-/** @type {import('./hook').Hook} */
-module.exports = {
+export default {
     afterLoad({ sourceFiles }) {
         const postActions = [];
         // 提取 example
@@ -66,19 +82,20 @@ module.exports = {
             const sourceName = sourceFile.getBaseNameWithoutExtension();
             const comments = sourceFile.getDescendantsOfKind(SyntaxKind.JSDocTag);
             const exampleTags = comments.filter((t) => t.getTagName() === 'example');
-            const pendingTextChangeAppliers = [];
+            const pendingTextChangeAppliers: { span: ts.TextSpan; newText: () => string }[] = [];
             for (const exampleTag of exampleTags) {
                 const jsdoc = exampleTag.getParentIfKindOrThrow(SyntaxKind.JSDoc);
                 const lineOffset = exampleTag.getStartLineNumber() - jsdoc.getStartLineNumber();
                 const exampleTagIdentifier = exampleTag.getTagNameNode();
                 const exampleCommentStartCol = exampleTagIdentifier.getEnd() - exampleTag.getStartLinePos();
                 const commentLines = jsdoc.getText().split('\n').slice(lineOffset);
-                const commentFirstLine = commentLines.shift().slice(exampleCommentStartCol);
+                const commentFirstLine = commentLines[0].slice(exampleCommentStartCol);
+                commentLines.shift();
                 let exampleName = commentFirstLine.trim();
                 const commentLinesWithoutStars = commentLines.map((l) => l.replace(/^\s*\*(?: )?/, ''));
-                const exampleParent = exampleTag.getParentWhileKind(SyntaxKind.JSDoc).getParent();
+                const exampleParent = exampleTag.getParentWhileKindOrThrow(SyntaxKind.JSDoc).getParent();
                 const exampleParentSymbol = exampleParent.getSymbol();
-                const sourceFileSymbol = sourceFile.getSymbol();
+                const sourceFileSymbol = sourceFile.getSymbolOrThrow();
                 let examplePath = '';
                 if (exampleParentSymbol) {
                     examplePath = exampleParentSymbol
@@ -95,7 +112,7 @@ module.exports = {
                         break;
                     }
                 }
-                let exampleContent = exampleTag.getCommentText().split('\n').slice(1).join('\n');
+                let exampleContent = (exampleTag.getCommentText() ?? '').split('\n').slice(1).join('\n');
                 const exampleVersions = examples[exampleName] ?? (examples[exampleName] = []);
                 if (commentLinesWithoutStars.length > 1 && commentLinesWithoutStars[0].startsWith('```')) {
                     const codeBlockEnd = commentLinesWithoutStars.indexOf('```', 1);
@@ -180,19 +197,20 @@ module.exports = {
             '@seeExample'
         ]);
         tsdocApplication.renderer.on('beginRender', () => {
-            const oldContextFactory = tsdocApplication.renderer.theme.getRenderContext;
-            tsdocApplication.renderer.theme.getRenderContext = function (...args) {
-                /** @type {import('typedoc').DefaultThemeRenderContext} */
+            const defaultTheme = tsdocApplication.renderer.theme as DefaultTheme;
+            // eslint-disable-next-line @typescript-eslint/unbound-method
+            const oldContextFactory = defaultTheme.getRenderContext;
+            defaultTheme.getRenderContext = function (...args) {
                 const renderContext = oldContextFactory.call(this, ...args);
-                const exampleTagName = renderContext.internationalization.translateTagName('@example');
                 const oldCommentTagsRender = renderContext.commentTags;
                 renderContext.commentTags = (props) => {
                     const jsx = oldCommentTagsRender(props);
-                    /** @type {JSX.Element[]} */
-                    const exampleTags = findJSXElement(
-                        jsx,
-                        (el) => typeof el === 'object' && el.props?.class?.includes(`tsd-tag-${exampleTagName}`)
-                    );
+                    const exampleTags = findJSXElement(jsx, (el): el is JSX.Element => {
+                        if (typeof el === 'object' && el.props) {
+                            return (el.props as { class?: string }).class?.includes(`tsd-tag-example`) ?? false;
+                        }
+                        return false;
+                    });
                     for (const exampleTag of exampleTags) {
                         const summaryTag = JSX.createElement('summary', null, exampleTag.children[0]);
                         const detailsTag = JSX.createElement('details', null, [
@@ -208,15 +226,18 @@ module.exports = {
         });
     },
     afterConvert({ tsdocProject }) {
+        const allReflections = Object.values(tsdocProject.reflections);
+        const reflAndSymbolIdMap = allReflections
+            .map((refl) => [refl, tsdocProject.getSymbolIdFromReflection(refl)] as const)
+            .filter((e): e is [Reflection, ReflectionSymbolId] => e[1] !== undefined);
         const exampleRefls = [];
         // 添加 example 页面
         const exampleParentRef = new DocumentReflection('示例', tsdocProject, [], { title: '示例' });
-        exampleParentRef._alias = 'examples';
-        tsdocProject.registerReflection(exampleParentRef);
+        tsdocProject.registerReflection(exampleParentRef, undefined, undefined);
         tsdocProject.addChild(exampleParentRef);
         for (const exampleName of Object.keys(examples).sort()) {
             const exampleVersions = examples[exampleName];
-            const content = [];
+            const content: CommentDisplayPart[] = [];
             for (const exampleVersion of exampleVersions) {
                 if (exampleVersions.length > 1) {
                     content.push({
@@ -239,9 +260,6 @@ module.exports = {
                 });
                 if (exampleVersion.sources.length > 1) {
                     exampleVersion.sources.sort((a, b) => {
-                        if (a.length !== b.length) {
-                            return b.length - a.length;
-                        }
                         if (a.source !== b.source) {
                             return a.source.localeCompare(b.source);
                         }
@@ -249,20 +267,23 @@ module.exports = {
                     });
                 }
                 for (const source of exampleVersion.sources) {
-                    const reflSymbolId = [...tsdocProject.symbolToReflectionIdMap.keys()].find(
-                        (k) => k.fileName === source.fileName && k.qualifiedName === source.path
+                    const reflAndSymbolId = reflAndSymbolIdMap.find(
+                        ([, symbolId]) =>
+                            symbolId.fileName === source.fileName && symbolId.qualifiedName === source.path
                     );
-                    const sourceRef = tsdocProject.getReflectionFromSymbolId(reflSymbolId);
-                    content.push({
-                        kind: 'text',
-                        text: '\n- '
-                    });
-                    content.push({
-                        kind: 'inline-tag',
-                        tag: '@link',
-                        text: source.path ? `${source.source} / ${source.path}` : source.source,
-                        target: sourceRef
-                    });
+                    if (reflAndSymbolId) {
+                        const sourceRef = tsdocProject.getReflectionFromSymbolId(reflAndSymbolId[1]);
+                        content.push({
+                            kind: 'text',
+                            text: '\n- '
+                        });
+                        content.push({
+                            kind: 'inline-tag',
+                            tag: '@link',
+                            text: source.path ? `${source.source} / ${source.path}` : source.source,
+                            target: sourceRef
+                        });
+                    }
                 }
                 content.push({
                     kind: 'text',
@@ -270,7 +291,7 @@ module.exports = {
                 });
             }
             const docRef = new DocumentReflection(exampleName, exampleParentRef, content, { title: exampleName });
-            tsdocProject.registerReflection(docRef);
+            tsdocProject.registerReflection(docRef, undefined, undefined);
             exampleRefls.push([exampleName, docRef, exampleVersions]);
             exampleParentRef.addChild(docRef);
         }
@@ -288,7 +309,7 @@ module.exports = {
                         if (relatedExample) {
                             commentTag.tag = '@example';
                             commentTag.name = exampleName;
-                            const replacement = [
+                            const replacement: CommentDisplayPart[] = [
                                 {
                                     kind: 'code',
                                     text: relatedExample.content
@@ -324,4 +345,4 @@ module.exports = {
             }
         }
     }
-};
+} as Hook;
