@@ -1,8 +1,8 @@
-const { SyntaxKind, ts } = require('ts-morph');
-const { resolve: resolvePath, relative: relativePath, sep: pathSep } = require('path');
-const { sep: pathSepPosix } = require('path/posix');
-const { mkdirSync, writeFileSync, existsSync, readFileSync } = require('fs');
-const { translatingPath, translatedPath } = require('./utils');
+import { existsSync, mkdirSync, readFileSync, writeFileSync } from 'fs';
+import { sep as pathSep, relative as relativePath, resolve as resolvePath } from 'path';
+import { sep as pathSepPosix } from 'path/posix';
+import { ExportGetableNode, SourceFile, SyntaxKind, ts } from 'ts-morph';
+import { translatedPath, translatingPath } from './utils.js';
 
 const SkippedTopLevelSyntaxKinds = [
     SyntaxKind.EndOfFileToken,
@@ -21,7 +21,7 @@ const SyntaxKindToCategory = new Map([
     [SyntaxKind.ExportDeclaration, 'types']
 ]);
 
-function getSourceFilePieceDirectory(sourceFile) {
+function getSourceFilePieceDirectory(sourceFile: SourceFile) {
     return resolvePath(
         translatingPath,
         relativePath(translatedPath, sourceFile.getDirectoryPath()),
@@ -29,7 +29,7 @@ function getSourceFilePieceDirectory(sourceFile) {
     );
 }
 
-function asRelativeModulePath(from, to) {
+function asRelativeModulePath(from: string, to: string) {
     const relativePathPlatform = relativePath(from, to);
     let relativePathPosix = relativePathPlatform.replaceAll(pathSep, pathSepPosix);
     if (!relativePathPosix.startsWith('.')) {
@@ -43,15 +43,30 @@ const ImportPrompt = '/* IMPORT */';
 const ExportPrompt = '/* EXPORT */';
 const PrivatePrompt = '/* PRIVATE */';
 
-/**
- * @param {import('ts-morph').SourceFile} sourceFile
- */
-function split(sourceFile) {
+export interface ExtractedPiece {
+    generated?: never;
+    start: number;
+    end: number;
+    path: string;
+    imports?: string;
+    exports?: string;
+}
+
+export interface GeneratedPiece {
+    generated: true;
+    start: -1;
+    path: string;
+    content: string;
+}
+
+export type Piece = ExtractedPiece | GeneratedPiece;
+
+export function split(sourceFile: SourceFile) {
     const pieceDirectory = getSourceFilePieceDirectory(sourceFile);
-    const indexExports = [];
-    const pieceExports = [];
-    const piecePathList = [];
-    const pieces = [];
+    const indexExports: string[] = [];
+    const pieceExports: { symbolName: string; piecePath: string; isExported: boolean }[] = [];
+    const piecePathList: string[] = [];
+    const pieces: Piece[] = [];
     const packageDocumentationJSDoc = sourceFile
         .getDescendantsOfKind(SyntaxKind.JSDoc)
         .find((jsdoc) => jsdoc.getTags().some((tag) => tag.getTagName() === 'packageDocumentation'));
@@ -63,7 +78,7 @@ function split(sourceFile) {
         });
     }
     const sourceScopeSymbols = new Set(
-        sourceFile.getSymbolsInScope(ts.SymbolFlags.ModuleMember).map((e) => e.getExportSymbol() || e)
+        sourceFile.getSymbolsInScope(ts.SymbolFlags.ModuleMember).map((e) => e.getExportSymbol())
     );
     sourceFile.forEachChild((node) => {
         if (SkippedTopLevelSyntaxKinds.includes(node.getKind())) return;
@@ -71,19 +86,18 @@ function split(sourceFile) {
             indexExports.push(node.getText());
             return;
         }
-        /** @type {import('ts-morph').Symbol[]} */
-        let includedSymbols = node
+        const includedSymbols = node
             .getDescendants()
             .map((e) => e.getSymbol())
             .filter((e) => e !== undefined);
         const scopedSymbols = includedSymbols.filter((e) => sourceScopeSymbols.has(e));
         let symbol = node.getSymbol();
         if (!symbol) {
+            if (includedSymbols.length === 0) {
+                console.log(`Cannot find symbol for node: ${node.getKindName()}`);
+                return;
+            }
             symbol = includedSymbols[0];
-        }
-        if (!symbol) {
-            console.log(`Cannot find symbol for node: ${node.getKindName()}`);
-            return;
         }
         const symbolName = symbol.getName();
         const category = SyntaxKindToCategory.get(node.getKind());
@@ -94,7 +108,7 @@ function split(sourceFile) {
         let piecePath = resolvePath(pieceDirectory, category, `${symbolName}.d.ts`);
         let pieceIndex = 1;
         while (piecePathList.includes(piecePath.toLowerCase())) {
-            piecePath = resolvePath(pieceDirectory, category, `${symbolName}-${pieceIndex}.d.ts`);
+            piecePath = resolvePath(pieceDirectory, category, `${symbolName}-${String(pieceIndex)}.d.ts`);
             pieceIndex++;
         }
         piecePathList.push(piecePath.toLowerCase());
@@ -102,8 +116,9 @@ function split(sourceFile) {
             .getChildrenOfKind(SyntaxKind.JSDoc)
             .filter((jsdoc) => jsdoc.getStart() !== packageDocumentationJSDoc?.getStart());
         let pieceStart = node.getStart(false);
-        if (jsdocList.length > 0) {
-            pieceStart = jsdocList.pop().getStart();
+        const firstJSDoc = jsdocList.pop();
+        if (firstJSDoc) {
+            pieceStart = firstJSDoc.getStart();
         }
 
         const importSymbols = [...new Set(scopedSymbols)]
@@ -123,15 +138,16 @@ function split(sourceFile) {
                         fromName: decl.getSymbolOrThrow().getName(),
                         toName: scopedSymbol.getName(),
                         sourceFile: resolvePath(getSourceFilePieceDirectory(decl.getSourceFile()), 'index.d.ts')
-                    }))[0];
+                    }))
+                    .shift();
             })
             .filter((e) => e !== undefined);
         importSymbols.sort((a, b) => (a.toName > b.toName ? 1 : a.toName < b.toName ? -1 : 0));
-        const isExported = node.hasExportKeyword?.() ?? false;
+        const isExported = (node as Partial<ExportGetableNode>).hasExportKeyword?.() ?? false;
 
-        const importGroupedByFile = {};
+        const importGroupedByFile: Record<string, string[]> = {};
         importSymbols.forEach((e) => {
-            let group = importGroupedByFile[e.sourceFile];
+            let group = importGroupedByFile[e.sourceFile] as string[] | undefined;
             if (!group) {
                 importGroupedByFile[e.sourceFile] = group = [];
             }
@@ -144,7 +160,7 @@ function split(sourceFile) {
             return `import { ${imports.join(', ')} } from '${fileNameRelative}';`;
         });
 
-        pieceExports.push([symbolName, piecePath, isExported]);
+        pieceExports.push({ symbolName, piecePath, isExported });
         pieces.push({
             start: pieceStart,
             end: node.getEnd(),
@@ -156,7 +172,7 @@ function split(sourceFile) {
 
     const sourceIndexPieceFile = resolvePath(pieceDirectory, 'index.d.ts');
     const sourceImportDeclarations = sourceFile.getImportDeclarations();
-    const indexImportStatements = [];
+    const indexImportStatements: string[] = [];
     const indexExportStatements = [...indexExports];
     sourceImportDeclarations.forEach((e) => {
         const importModulePathRelative = asRelativeModulePath(
@@ -164,35 +180,39 @@ function split(sourceFile) {
             getSourceFilePieceDirectory(e.getModuleSpecifierSourceFileOrThrow())
         );
         const importClause = e.getImportClause();
+        if (!importClause) {
+            throw new Error(`Cannot find import clause for import statement: ${e.print()}`);
+        }
         const importedIdentifiers = [];
         const namespaceImport = importClause.getNamespaceImport();
         if (namespaceImport) {
             importedIdentifiers.push(namespaceImport);
         } else {
             importClause.getNamedImports().forEach((spec) => {
-                importedIdentifiers.push(spec.getAliasNode() || spec.getNameNode());
+                importedIdentifiers.push(spec.getAliasNode() ?? spec.getNameNode());
             });
         }
         const importedIdentifierNames = importedIdentifiers.map((e) => e.getText());
         indexImportStatements.push(`import ${importClause.getText()} from '${importModulePathRelative}';`);
         indexExportStatements.push(`${PrivatePrompt} export { ${importedIdentifierNames.join(', ')} };`);
     });
-    pieceExports.forEach(([symbolName, piecePath, isExported]) => {
+    pieceExports.forEach(({ symbolName, piecePath, isExported }) => {
         const piecePathRelative = asRelativeModulePath(pieceDirectory, piecePath);
         const prefix = isExported ? '' : `${PrivatePrompt} `;
         indexExportStatements.push(`${prefix}export { ${symbolName} } from '${piecePathRelative}';`);
     });
     const sourceIndexPiece = {
         generated: true,
+        start: -1,
         path: sourceIndexPieceFile,
         content: [...indexImportStatements, ...indexExportStatements].join('\n')
-    };
+    } as const;
     pieces.sort((a, b) => b.start - a.start);
     pieces.unshift(sourceIndexPiece);
     return pieces;
 }
 
-function writePiece(sourceFile, piece) {
+export function writePiece(sourceFile: SourceFile, piece: Piece) {
     mkdirSync(resolvePath(piece.path, '..'), { recursive: true });
     if (piece.generated) {
         writeFileSync(piece.path, piece.content);
@@ -208,7 +228,7 @@ function writePiece(sourceFile, piece) {
     }
 }
 
-function replacePieces(sourceFile, pieces) {
+export function replacePieces(sourceFile: SourceFile, pieces: Piece[]) {
     let sourceFileText = sourceFile.getFullText();
     let writtenCount = 0;
     pieces.forEach((piece) => {
@@ -225,5 +245,3 @@ function replacePieces(sourceFile, pieces) {
         sourceFile.replaceWithText(sourceFileText);
     }
 }
-
-module.exports = { split, writePiece, replacePieces };
